Major structuring overhaul, and functioning search feature
