+++ src/components/Welcome.js
@@ -1,0 +1,14 @@
+const Welcome = (props) => {
+    return (
+        <>
+            <h4>*⬇ START OF WELCOME COMPONENT ⬇*</h4>
+            <div className="welcome">
+                <h2>Welcome to our website!</h2>
+                <p>Use the search component below!</p>
+            </div>
+            <h4>*⬆ END OF WELCOME COMPONENT ⬆*</h4>
+        </>
+    )
+}
+
+export default Welcome
